Docs and code refactoring.
